fixing getters and setter for the year attribute to the int type

+++ bookstore/src/main/java/com/plakolli/bookstore/domain/Book.java
@@ -59,11 +59,11 @@ public class Book {
 		this.isbn = isbn;
 	}
 
-	public Long getYear() {
+	public int getYear() {
 		return year;
 	}
 
-	public void setYear(Long year) {
+	public void setYear(int year) {
 		this.year = year;
 	}
